+++ tomcat/src/main/java/nextstep/jwp/controller/AbstractController.java
@@ -1,8 +1,8 @@
 package nextstep.jwp.controller;
 
-import org.apache.coyote.http11.HttpMethod;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpMethod;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
 public abstract class AbstractController implements Controller {
