+++ tomcat/src/main/java/nextstep/jwp/controller/Controller.java
@@ -1,7 +1,7 @@
 package nextstep.jwp.controller;
 
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
 public interface Controller {
