+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestUri.java
@@ -1,4 +1,6 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
+
+import org.apache.coyote.http11.MediaType;
 
 public class RequestUri {
